(feat) added portfolio analytics route, updated StockQuote schema to use Decimal, added unit testests, API tests

+++ src/atlas_api/schemas/stock.py
@@ -1,15 +1,16 @@
 #!/usr/bin/env python3
 
 from pydantic import BaseModel
+from decimal import Decimal
 
 
 class StockQuote(BaseModel):
     symbol: str 
-    current_price: float
-    price_change: float
-    percent_change: float
-    high_price: float
-    low_price: float
-    open_price: float
-    previous_close_price: float
+    current_price: Decimal
+    price_change: Decimal
+    percent_change: Decimal
+    high_price: Decimal
+    low_price: Decimal
+    open_price: Decimal
+    previous_close_price: Decimal
     timestamp: int
